requirements.txt
determined-cli
